Secondary call info displayed after far end disconnected
During conference call, another MO/MT call is established and
it is put to HOLD. So conference call is ACTIVE and manage
conference screen is opened at this sceanrio if the HOLD call is
disconnected, secondary info is still visible. check if manage
conference fragment is opened when the any secondary call info
need to be updated.

CRs-Fixed: 1001701, 963990
Change-Id: I887a93adf7a20bada2ee25f14586f79520f83812