InCallUI:Send broadcast to SystemUI to update StatusBar
When user press home button or switch to other app during calling,
it will send a broadcast which contains the call's information to
the Status Bar for displaying the calling status. And user can press
the Status Bar to return to calling interface at once or change
audio mode.

Change-Id: I252de604503bdf52c0e1824409c853ed996033fe
CRs-Fixed: 985477