Call notification not displayed at SIM query UI
When the SIM slection query UI is launched from InCallActivty and
when the device receives incoming call, the current state of call ui
considered as IN_CALL, due to this notification is not displaying.
Audio is not heard due to the mismatch in the foreground call. check
whether SIM selection ui is diplayed.

Change-Id: Ic283825bc58cdc7534b4ab71b7ec93770b7d1d1a
CRs-Fixed: 1014066